+++ scripts/main.js
@@ -7,21 +7,8 @@ const addListenersTrackedMovies = require("./trackedMovies/addListenersTrackedMo
 const addListenersCard = require("./addListenersCards");
 
 auth.init()
-const dataManager = require("./util/datamanager.js")
 
 
-$(document).ready(function () {
-    $(".modal").modal({
-        dismissable: true,
-        complete: function() {
-            $(".movie-rating__item").removeClass("movie-rating__item--starred")
-            let targetId = $(".rated").attr("id")
-            const rating = parseInt(targetId.split("_")[1]);
-            const stuff = [parseInt($("#rating__modal").attr("data-firebaseId")), parseInt($("#rating__modal").attr("data-movieId"))]
-            dataManager.firebasePUT($("#rating__modal").attr("data-firebaseId"),{movidId: parseInt($("#rating__modal").attr("data-movieId")), rating: rating}).then(r=> {})
-        }
-    });
-});
 
 $(".movie-rating__item") //highlight stars on hover
     .hover(function () {
